move dex stats to chain context
