+++ src/components/Home/Stats/AmbientStats.tsx
@@ -1,18 +1,6 @@
-import { CrocEnv } from '@crocswap-libs/sdk';
-import { useContext, useEffect, useMemo, useState } from 'react';
+import { useContext } from 'react';
 import { Fade } from 'react-reveal';
-import {
-    GCGO_BLAST_URL,
-    GCGO_ETHEREUM_URL,
-    GCGO_SCROLL_URL,
-} from '../../../ambient-utils/constants';
-import {
-    getChainStats,
-    getFormattedNumber,
-} from '../../../ambient-utils/dataLayer';
-import { CachedDataContext } from '../../../contexts/CachedDataContext';
-import { CrocEnvContext } from '../../../contexts/CrocEnvContext';
-import { TokenContext } from '../../../contexts/TokenContext';
+import { ChainDataContext } from '../../../contexts';
 import {
     HomeContent,
     HomeTitle,
@@ -29,6 +17,7 @@ interface StatCardProps {
 
 function StatCard(props: StatCardProps) {
     const { title, value } = props;
+
     const ariaDescription = `${title} is ${value}`;
     return (
         <StatCardContainer
@@ -54,205 +43,8 @@ function StatCard(props: StatCardProps) {
 }
 
 export default function Stats() {
-    const { mainnetProvider, scrollProvider, blastProvider } =
-        useContext(CrocEnvContext);
-    const { tokens } = useContext(TokenContext);
-    const allDefaultTokens = tokens.allDefaultTokens;
-
-    const mainnetCrocEnv = useMemo(
-        () =>
-            mainnetProvider
-                ? new CrocEnv(mainnetProvider, undefined)
-                : undefined,
-        [mainnetProvider !== undefined],
-    );
-
-    const scrollCrocEnv = useMemo(
-        () =>
-            scrollProvider ? new CrocEnv(scrollProvider, undefined) : undefined,
-        [scrollProvider !== undefined],
-    );
-
-    const blastCrocEnv = useMemo(
-        () =>
-            blastProvider ? new CrocEnv(blastProvider, undefined) : undefined,
-        [blastProvider !== undefined],
-    );
-
-    const { cachedFetchTokenPrice } = useContext(CachedDataContext);
-
-    const [totalTvlString, setTotalTvlString] = useState<string | undefined>();
-    const [totalVolumeString, setTotalVolumeString] = useState<
-        string | undefined
-    >();
-    const [totalFeesString, setTotalFeesString] = useState<
-        string | undefined
-    >();
-
-    useEffect(() => {
-        if (
-            mainnetCrocEnv !== undefined &&
-            scrollCrocEnv !== undefined &&
-            blastCrocEnv !== undefined &&
-            allDefaultTokens.length > 0
-        ) {
-            let tvlTotalUsd = 0,
-                volumeTotalUsd = 0,
-                feesTotalUsd = 0;
-
-            const numChainsToAggregate = 3;
-            let resultsReceived = 0;
-
-            getChainStats(
-                'cumulative',
-                '0x1',
-                mainnetCrocEnv,
-                GCGO_ETHEREUM_URL,
-                cachedFetchTokenPrice,
-                10,
-                allDefaultTokens,
-            ).then((dexStats) => {
-                if (!dexStats) {
-                    return;
-                }
-                tvlTotalUsd += dexStats.tvlTotalUsd;
-                volumeTotalUsd += dexStats.volumeTotalUsd;
-                feesTotalUsd += dexStats.feesTotalUsd;
-
-                resultsReceived += 1;
-                // IS_LOCAL_ENV &&
-                //     console.log(
-                //         'mainnet cumulative vol: ',
-                //         dexStats.volumeTotalUsd.toLocaleString(),
-                //     );
-
-                if (resultsReceived === numChainsToAggregate) {
-                    setTotalTvlString(
-                        getFormattedNumber({
-                            value: tvlTotalUsd,
-                            prefix: '$',
-                            isTvl: true,
-                            mantissa: 1,
-                        }),
-                    );
-                    setTotalVolumeString(
-                        getFormattedNumber({
-                            value: volumeTotalUsd,
-                            prefix: '$',
-                            mantissa: 1,
-                        }),
-                    );
-                    setTotalFeesString(
-                        getFormattedNumber({
-                            value: feesTotalUsd,
-                            prefix: '$',
-                            mantissa: 1,
-                        }),
-                    );
-                }
-            });
-
-            getChainStats(
-                'cumulative',
-                '0x82750',
-                scrollCrocEnv,
-                GCGO_SCROLL_URL,
-                cachedFetchTokenPrice,
-                20,
-                allDefaultTokens,
-            ).then((dexStats) => {
-                if (!dexStats) {
-                    return;
-                }
-                tvlTotalUsd += dexStats.tvlTotalUsd;
-                volumeTotalUsd += dexStats.volumeTotalUsd;
-                feesTotalUsd += dexStats.feesTotalUsd;
-                resultsReceived += 1;
-                // IS_LOCAL_ENV &&
-                // console.log(
-                //     'scroll cumulative vol: ',
-                //     dexStats.volumeTotalUsd.toLocaleString(),
-                // );
-
-                if (resultsReceived === numChainsToAggregate) {
-                    setTotalTvlString(
-                        getFormattedNumber({
-                            value: tvlTotalUsd,
-                            prefix: '$',
-                            isTvl: true,
-                            mantissa: 1,
-                        }),
-                    );
-                    setTotalVolumeString(
-                        getFormattedNumber({
-                            value: volumeTotalUsd,
-                            prefix: '$',
-                            mantissa: 1,
-                        }),
-                    );
-                    setTotalFeesString(
-                        getFormattedNumber({
-                            value: feesTotalUsd,
-                            prefix: '$',
-                            mantissa: 1,
-                        }),
-                    );
-                }
-            });
-
-            getChainStats(
-                'cumulative',
-                '0x13e31',
-                blastCrocEnv,
-                GCGO_BLAST_URL,
-                cachedFetchTokenPrice,
-                10,
-                allDefaultTokens,
-            ).then((dexStats) => {
-                if (!dexStats) {
-                    return;
-                }
-                tvlTotalUsd += dexStats.tvlTotalUsd;
-                volumeTotalUsd += dexStats.volumeTotalUsd;
-                feesTotalUsd += dexStats.feesTotalUsd;
-                resultsReceived += 1;
-                // IS_LOCAL_ENV &&
-                //     console.log(
-                //         'blast cumulative vol: ',
-                //         dexStats.volumeTotalUsd.toLocaleString(),
-                //     );
-                if (resultsReceived === numChainsToAggregate) {
-                    setTotalTvlString(
-                        getFormattedNumber({
-                            value: tvlTotalUsd,
-                            prefix: '$',
-                            isTvl: true,
-                            mantissa: 1,
-                        }),
-                    );
-                    setTotalVolumeString(
-                        getFormattedNumber({
-                            value: volumeTotalUsd,
-                            prefix: '$',
-                            mantissa: 1,
-                        }),
-                    );
-                    setTotalFeesString(
-                        getFormattedNumber({
-                            value: feesTotalUsd,
-                            prefix: '$',
-                            mantissa: 1,
-                        }),
-                    );
-                }
-            });
-        }
-    }, [
-        mainnetCrocEnv !== undefined &&
-            scrollCrocEnv !== undefined &&
-            blastCrocEnv !== undefined &&
-            allDefaultTokens.length > 0,
-    ]);
+    const { totalTvlString, totalVolumeString, totalFeesString } =
+        useContext(ChainDataContext);
 
     const statCardData = [
         {
